crdbtest: move crdb rowlbk benchmarks
Move the remaining crdb rowblk benchmarks to crdb and unexport test
key generation functionality.
